project/requirements.txt
alembic==0.9.9
astroid==1.6.2
click==6.7
Flask==0.12.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
newsapi-python==0.0.2
psycopg2==2.7.4
pylint==1.8.3
python-dateutil==2.7.2
python-editor==1.0.3
requests==2.13.0
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
